requirements.txt
python-dotenv>=0.21.0
google-generativeai>=0.1.0
python-telegram-bot>=20.0
pymorphy2>=0.9